Move relative movement classes to package movement.relative
