+++ src/main/java/org/mastodon/mamut/feature/branch/movement/relative/BranchRelativeMovementFeature.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package org.mastodon.mamut.feature.branch.relativemovement;
+package org.mastodon.mamut.feature.branch.movement.relative;
 
 import org.mastodon.feature.Dimension;
 import org.mastodon.feature.Feature;
