+++ src/main/java/org/mastodon/mamut/feature/spot/movement/relative/SpotRelativeMovementFeature.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package org.mastodon.mamut.feature.spot.relativemovement;
+package org.mastodon.mamut.feature.spot.movement.relative;
 
 import org.mastodon.feature.Dimension;
 import org.mastodon.feature.Feature;
